+++ ios/engine/KMEI/KeymanEngine/Classes/Settings/LanguageSettingsViewController.swift
@@ -14,9 +14,11 @@ class LanguageSettingsViewController: UITableViewController {
   let language: Language
   private var userKeyboards: [String: Language] = [:]
   private var settingsArray = [[String: String]]()
+  private var keyboardRepository: KeyboardRepository?
 
-  public init(_ inLanguage: Language) {
+  public init(_ keyboardRepository: KeyboardRepository?, _ inLanguage: Language) {
     language = inLanguage
+    self.keyboardRepository = keyboardRepository
     super.init(nibName: nil, bundle: nil)
   }
   
@@ -288,7 +290,7 @@ class LanguageSettingsViewController: UITableViewController {
   func showAddLanguageKeyboard() {
     let button: UIButton? = (navigationController?.toolbar?.viewWithTag(toolbarButtonTag) as? UIButton)
     button?.isEnabled = false
-    let vc = LanguageSpecificViewController(Manager.shared.apiKeyboardRepository, language: language)
+    let vc = LanguageDetailViewController(keyboardRepository, language: language)
     vc.title = "Add new \(language.name) keyboard"
     navigationController?.pushViewController(vc, animated: true)
   }
@@ -315,7 +317,7 @@ class LanguageSettingsViewController: UITableViewController {
     let userData = Storage.active.userDefaults
 
     // If user defaults for keyboards list does not exist, do nothing.
-    guard var globalUserKeyboards = userData.userKeyboards else {
+    guard let globalUserKeyboards = userData.userKeyboards else {
       log.error("no keyboards in the global keyboards list!")
       return nil
     }